Extract integration-common for shared types
