+++ test/integration-test/Cargo.toml
@@ -14,6 +14,7 @@ assert_matches = { workspace = true }
 aya = { path = "../../aya", version = "^0.13.1", default-features = false }
 aya-log = { path = "../../aya-log", version = "^0.2.1", default-features = false }
 aya-obj = { path = "../../aya-obj", version = "^0.2.1", default-features = false }
+integration-common = { path = "../integration-common", features = ["user"] }
 env_logger = { workspace = true }
 epoll = { workspace = true }
 futures = { workspace = true, features = ["std"] }
